Update Maven support for 1.8
I rewrote the POM file to include the Spigot repositories, which includes the Bukkit and Spigot APIs for 1.8. This will give you a more modern API to work with within your project.

Additionally, I changed the Metrics dependency to R8-SNAPSHOT, which is the only Metrics version that supports Spigot 1.8.

Otherwise, I just organized it more tidily per documentation on the Bukkit Wiki to try to prevent issues farther into the future.